fix: revert structure
