+++ src/lib/pages/proposal-details/components/index.ts
@@ -1,2 +1,4 @@
-export * from "./proposal-details-body";
+export * from "./proposal-overview";
+export * from "./proposal-top";
+export * from "./vote-details";
 export * from "./InvalidProposal";
